fixed fatal notify bugs
greatly reduced API request rate,
made semester and years parameters to checkNotify
